Extract the inserts in table `results_propagate` out of the `before_delete_items_items`, and put it at the beginning of the process of `results propagation`.
Problem: we have a case in production where 50k inserts are done in the trigger, and this makes the Lambda killed. One insert is to be made for each entry in table `results` with a specified `item_id`.

Solution: instead of doing the insert in the trigger, we add the `item_id` in question in a new table (`results_propagate_items`). Then, when we do the `results propagation`, we first retrieve the `item_id` in the new table, and do the inserts at this moment.

We do one transaction for each `item_id`, because we know it can take a long time, and we want to keep transactions as small as possible.

Signed-off-by: Geoffrey Huck <geoffrey@geot.fr>