`saveGrade`: remove `task_token` from the input. We get the parameters from either the `score_token`, or from `score`+`answer_token`.
Since we need the `localItemID` contained in the `score_token` to know which key to use to verify the signature of the `score_token`, the `GetUnsafeFromToken` function was added. `Unsafe` is present is the name to make it explicit that the signature wasn't verified at that point.

Once the tokens are unmarshalled, we store the parameters we need in `saveGradeRequestParsed`.`Converted`, so we don't need to worry about how the parameters were passed.

Signed-off-by: Geoffrey Huck <geoffrey@geot.fr>