`groupUpdate` when approval are strengthened.
Use the transition system that was developed instead of adding the entries in the database. For consistence.

"reinvite" didn't work, because we didn't add the entries in table `group_membership_changes`. Now it should be fine because the transition system takes care of it.

Two slight changes from before when the approval are strengthened:
- `leave_request` doesn't get re-invited nor stay in `leave_request`. They just leave. I think that's the expected behavior.

Signed-off-by: Geoffrey Huck <geoffrey@geot.fr>