Revert "itemActivityLogForItem: activity_type is passed as a string" as (by points):
"We need to optimize the sub-query on table answers because it takes a lot of time"
Right, but this change didn't help unfortunately.

"It takes a lot of time because the "order by" of the query is not indexable

The integer enum that was used before prevented the creation of the index"
That's not true. The integer enum does really prevent the usage of the whole index (`created_at` DESC,`item_id`,`participant_id`,`attempt_id` DESC,`type` DESC,`id`), but its trailing part (`created_at` DESC,`item_id`,`participant_id`,`attempt_id` DESC) was still usable. So the integer enum only prevented MySQL from using the index for ordering rows having the same (`created_id`, `item_id`, `participant_id`, `attempt_id`) which is a very rare case. That's not the cause of the slow down. Anyway, it can be fixed easily without rewriting everything and breaking the ordering logic.

"Since the sort on activity type is merely used as a tie-breaker, the order doesn't really matter"
That's not true. The order does matter. For events from different tables having the same timestamp, we know that result starting goes before submission of the answer and result validation goes after submission of the answer. The tests had been checked that carefully.

"We also beed activity_type to be consistent accross the query and its 3 subqueries because they all use the pagination system"
A string value is as consistent across 3 subqueries as an integer value. Also, a string requires much memory to to store results of subqueries and the union andtherefore slows the query down.

This reverts commit 8271d7a9ff224302a818f67f9826ae34c5351870.